requirements.txt
django
djangorestframework
drf-yasg
django-jazzmin
pillow